Make OutboundHandler protocol-dependant (InboundHandler)
