+++ server/src/main/java/org/opensearch/transport/ProtocolMessageHandler.java
@@ -24,4 +24,6 @@ public interface ProtocolMessageHandler {
         long slowLogThresholdMs,
         TransportMessageListener messageListener
     ) throws IOException;
+
+    public void setMessageListener(TransportMessageListener listener);
 }
