+++ server/src/main/java/org/opensearch/transport/ProtocolOutboundHandler.java
@@ -1,0 +1,51 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.transport;
+
+import org.opensearch.Version;
+import org.opensearch.cluster.node.DiscoveryNode;
+import org.opensearch.core.transport.TransportResponse;
+
+import java.io.IOException;
+import java.util.Set;
+
+abstract class ProtocolOutboundHandler {
+
+    abstract void sendRequest(
+        final DiscoveryNode node,
+        final TcpChannel channel,
+        final long requestId,
+        final String action,
+        final TransportRequest request,
+        final TransportRequestOptions options,
+        final Version channelVersion,
+        final boolean compressRequest,
+        final boolean isHandshake
+    ) throws IOException, TransportException;
+
+    abstract void sendResponse(
+        final Version nodeVersion,
+        final Set<String> features,
+        final TcpChannel channel,
+        final long requestId,
+        final String action,
+        final TransportResponse response,
+        final boolean compress,
+        final boolean isHandshake
+    ) throws IOException;
+
+    abstract void sendErrorResponse(
+        final Version nodeVersion,
+        final Set<String> features,
+        final TcpChannel channel,
+        final long requestId,
+        final String action,
+        final Exception error
+    ) throws IOException;
+}
